tried to implement repeat_ids
Problem is that it can be (5,5,5,3) for example (same image always repeating for x triplets) in a row. Also, it's not the the 1st/3rd image of the same triplet preceding it, it can be any image of any given triplet (i.e. ABC then "D" DEF then "A" GHI).